Use proper options and add some
Move previous arguments to getopt for processing.

Add a flag for excluding Qemus that don't have any RBDs. An example
of this is would be e.g. an io.* VM which was booted in Jewel era
and does not have any volumes attached.

Add a debug flag which allows outputting individual RBD counts (since
that seems to correlate nicely to the number that `ceph --features`
outputs.