CCCP-3802: Limit the exclude of RBD to VMs rather than servers
In the current version of the code, when a VM is detected does not have
an RBD and the flag exclude_qemu_without_rbd is set, the script omit to
check other VMs in the server. This fix allows to list all the VMs that
have RBD in a server that has a VM without rbd and the flag
exclude_qemu_without_rbd is set.